improved gradle configuration, rewrited and added severl unit tests
