--- provider_receiver/src/main/java/com/receiver/utils/Provider.java
@@ -1,14 +1,0 @@
-package com.receiver.utils;
-
-import org.apache.avro.generic.GenericRecord;
-import org.apache.hadoop.io.Text;
-import org.apache.hadoop.mapreduce.Mapper;
-
-/**
- * Created by ramizjon on 3/1/16.
- */
-public abstract class Provider {
-
-    public abstract boolean isOptional();
-    public abstract Mapper<?, ?, Void, GenericRecord> getMapper();
-}
