+++ unifier/src/main/java/com/unifier/nexusprovider/NexusProvider.java
@@ -1,0 +1,17 @@
+package com.unifier.nexusprovider;
+
+import com.unifier.mapreduce.AppContext;
+import com.unifier.utils.Provider;
+import lombok.Data;
+import org.apache.avro.generic.GenericRecord;
+import org.apache.hadoop.mapreduce.Mapper;
+
+@Data
+public class NexusProvider extends Provider {
+
+    private final boolean isOptional = false;
+
+    public Mapper<?, ?, Void, GenericRecord> getMapper() {
+        return new AppContext.NexusUserSegmentsMapperImpl();
+    }
+}
